Refactoring and cleanup of "module wizards"
