+++ packages/app/src/views/AddModule/index.tsx
@@ -4,7 +4,7 @@ import { ZodiacPaper } from "zodiac-ui-components"
 import { ModuleButton } from "./ModuleButton"
 import { useRootDispatch, useRootSelector } from "../../store"
 import { getModulesList } from "../../store/modules/selectors"
-import { ModuleModals } from "./wizards/ModuleModals"
+import { ModuleModals } from "./wizards/ModalWizards"
 import { ModuleType } from "../../store/modules/models"
 import { fetchPendingModules, setModuleAdded, setRealityModuleScreen } from "../../store/modules"
 import { useSafeAppsSDK } from "@gnosis.pm/safe-apps-react-sdk"
