+++ packages/app/src/views/AddModule/wizards/readme.md
@@ -2,12 +2,12 @@
 
 Wizards for setting up and attaching Zodiac Modules to a Safe. The wizards can either be displayed in a Modal or use the whole right side screen view.
 
-In both cases, the a Module Button for the new wizard must be added to the `AddModule/index.tsx` file.
+In both cases, the a `ModuleButton` for the new wizard must be added to the `AddModule/index.tsx` file.
 
 ## Modal
 
-Wizards can be displayed in a Modal using the AddModuleModal and adding the Modal to ModuleModals. An example of this is the Delay Module wizard.
+Wizards can be displayed in a Modal using the `AddModuleModal` component and adding the Modal to `ModalWizards` component. An example of this is the `DelayModule` wizard.
 
 ## Full view
 
-Wizards can also take up the whole right side view (man view). An example of this is the Reality Module wizard.
+Wizards can also take up the whole right side view (man view). An example of this is the `RealityModule` wizard.
